refactor : model, controller 폴더로 분리
