+++ src/main/java/lotto/model/WinningCalculator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.math.BigDecimal;
 import java.util.LinkedHashMap;
