+++ src/test/java/lotto/WinningLottoWithBonusTest.java
@@ -1,9 +1,10 @@
 package lotto;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
+import lotto.model.Lotto;
+import lotto.model.WinningLottoWithBonus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
